Enhancement of Explore Page
