+++ app/templates/components/explore/side-bar.hbs
@@ -10,7 +10,7 @@
   <div class="ui input">
     <Input
       @name="name"
-      @value={{this.event_name}}
+      @value={{this.name}}
       @type="text"
       @placeholder={{t "Enter Event Name"}} />
   </div>
